crud produto

+++ src/main/java/com/server/restauranteserver/beans/ProdutoBEAN.java
@@ -21,7 +21,7 @@ private byte[] foto;
 private String preparo;
 private String descricao;
 private String tipo;
-private int quantidade;
+private float quantidade;
 
 
     public int getCodigo() {
@@ -96,13 +96,15 @@ private int quantidade;
         this.tipo = tipo;
     }
 
-    public int getQuantidade() {
+    public float getQuantidade() {
         return quantidade;
     }
 
-    public void setQuantidade(int quantidade) {
+    public void setQuantidade(float quantidade) {
         this.quantidade = quantidade;
     }
+
+
